swapping to names queries

+++ src/test/java/uk/ac/vfb/geppetto/test/MultipleQueriesVFBQueryTest.java
@@ -181,12 +181,12 @@ public class MultipleQueriesVFBQueryTest
 		EList<RunnableQuery> runnableQueriesEMF = new BasicEList<RunnableQuery>();
 
 		RunnableQuery rqEMF1 = DatasourcesFactory.eINSTANCE.createRunnableQuery();
-		rqEMF1.setQueryPath(model.getQueries().get((int) avQ.get("neuronssynaptic")).getPath());
+		rqEMF1.setQueryPath(model.getQueries().get(avQ.get("neuronssynaptic")).getPath());
 		rqEMF1.setTargetVariablePath(variable1.getPath());
 		runnableQueriesEMF.add(rqEMF1);
 		
 		RunnableQuery rqEMF2 = DatasourcesFactory.eINSTANCE.createRunnableQuery();
-		rqEMF2.setQueryPath(model.getQueries().get((int) avQ.get("neuronssynaptic")).getPath());
+		rqEMF2.setQueryPath(model.getQueries().get(avQ.get("neuronssynaptic")).getPath());
 		rqEMF2.setTargetVariablePath(variable2.getPath());
 		runnableQueriesEMF.add(rqEMF2);
